lurk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lurk"
version = "0.1.0"
description = "Context broker for AI tools — the missing memory layer"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mcp = ["mcp[cli]>=1.0"]
http = ["uvicorn>=0.27", "starlette>=0.36"]
llm = ["httpx>=0.27"]
all = ["lurk[mcp,http,llm]"]

[project.scripts]
lurk = "lurk.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/lurk"]
